Move light-specific applicators out of plAnimation.
This matches how the plSoundVolumeApplicator is handled. More
importantly, it removes the need for plAnimation to link against
plGLight.
