+++ Sources/Plasma/PubUtilLib/plAnimation/plScalarChannel.h
@@ -308,55 +308,4 @@ public:
     GETINTERFACE_ANY( plScalarChannelApplicator, plAGApplicator );
 };
 
-class plSpotInnerApplicator : public plAGApplicator
-{
-protected:
-    virtual void IApply(const plAGModifier *mod, double time);
-
-public:
-    CLASSNAME_REGISTER( plSpotInnerApplicator );
-    GETINTERFACE_ANY( plSpotInnerApplicator, plAGApplicator );
-};
-
-class plSpotOuterApplicator : public plAGApplicator
-{
-protected:
-    virtual void IApply(const plAGModifier *mod, double time);
-
-public:
-    CLASSNAME_REGISTER( plSpotOuterApplicator );
-    GETINTERFACE_ANY( plSpotOuterApplicator, plAGApplicator );
-};
-
-class plOmniApplicator : public plAGApplicator
-{
-protected:
-    virtual void IApply(const plAGModifier *mod, double time);
-
-public:
-    CLASSNAME_REGISTER( plOmniApplicator );
-    GETINTERFACE_ANY( plOmniApplicator, plAGApplicator );
-};
-
-class plOmniSqApplicator : public plAGApplicator
-{
-protected:
-    virtual void IApply(const plAGModifier *mod, double time);
-
-public:
-    CLASSNAME_REGISTER( plOmniSqApplicator );
-    GETINTERFACE_ANY( plOmniSqApplicator, plAGApplicator );
-};
-
-class plOmniCutoffApplicator : public plAGApplicator
-{
-protected:
-    virtual void IApply(const plAGModifier *mod, double time);
-
-public:
-    CLASSNAME_REGISTER( plOmniCutoffApplicator );
-    GETINTERFACE_ANY( plOmniCutoffApplicator, plAGApplicator );
-};
-
-
 #endif
